+++ SurfaceTopography/IO/__init__.py
@@ -47,6 +47,7 @@ from .NPY import NPYReader
 from .PS import PSReader
 from .OPD import OPDReader
 from .OPDx import OPDxReader
+from .SUR import SURReader
 from .VK import VKReader
 from .ZON import ZONReader
 
@@ -72,6 +73,7 @@ readers = [
     H5Reader,
     NPYReader,
     PSReader,
+    SURReader,
     VKReader,
     ZONReader,
     AL3DReader,